Add files via upload
This code draws a pie chart that has multiple segments. The size of the segments depend on the numbers entered at the keyboard. Since a circle has just 360 degrees, I limited the number of sub-divisions to 360. So basically, the code will not allow you to enter more than 360 numbers at a time.